refactor into service for easier testing
