+++ src/Exception/PreloadWriteException.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+namespace CakePreloader\Exception;
+
+use Cake\Core\Exception\CakeException;
+
+/**
+ * Thrown when a preloader file fails to be written
+ */
+class PreloadWriteException extends CakeException
+{
+
+}
